refactor(Lotto): Lotto 파일 패키지 변경
* Lotto 자바 파일을 domain 패키지로 이동
