+++ src/main/java/lotto/controller/LottoGameController.java
@@ -1,6 +1,6 @@
 package lotto.controller;
 
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.domain.*;
 import lotto.view.InputView;
 import lotto.view.OutputView;
