+++ src/main/java/lotto/domain/Lotto.java
@@ -1,11 +1,8 @@
-package lotto;
-
-import lotto.domain.BonusNumber;
+package lotto.domain;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 public class Lotto {
     private final List<Integer> numbers;
