+++ src/main/java/lotto/domain/LottoOwner.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.Lotto;
-
 import java.util.List;
 
 public class LottoOwner {
