+++ src/test/java/lotto/WinningNumberTest.java
@@ -1,15 +1,14 @@
 package lotto;
 
 import lotto.domain.BonusNumber;
+import lotto.domain.Lotto;
 import lotto.domain.Rank;
 import lotto.domain.WinningNumber;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
